Reorg libs/utils
