+++ src/utils/updateSprintData.ts
@@ -1,24 +1,4 @@
-import { I18n } from "@aws-amplify/core";
 import { WS } from "../libs/websocket";
-
-/**
- * @constant
- */
-export const steps = [
-  {
-    selector: ".first-step-arena",
-    content: `${I18n.get("arenaFirst")}`,
-  },
-  {
-    selector: ".second-step-arena",
-    content: `${I18n.get("arenaSecond")}`,
-  },
-  {
-    selector: ".third-step-arena",
-    content: `${I18n.get("arenaThird")}`,
-  },
-];
-
 /**
  * This function sends sprint data to the server.
  * @param sprintData data to be sent over to the sprint
